MVC
We created controllers for all the request handling methods. By first creating a directory 'controllers' then a file for blog related controllers 'blogController' then created a funtion for each function previously passed into the request handling method. Then export the functions and imported them to the 'blogRoutes' file and passed the functions into the request handling methods (this time only function names not the entire code per function) and scoped the views
