+++ routes/blogRoutes.js
@@ -4,7 +4,10 @@
 const express = require('express');
 // creating an express router
 const router = express.Router(); //create a new instance of a Router object
-const Blog = require('../models/blog');
+
+//importing blogController
+const blogController = require('../controllers/blogController');
+
 
 // blog routes
 // replacing the 'app' instance with the 'router' instance (ex: app.get => router.get ...)
@@ -20,51 +23,31 @@ const Blog = require('../models/blog');
 // });
 
 //after scoped out 
-router.get('/', (req,res) => {
-    Blog.find().sort({ createdAt : -1 })
-        .then(result => {
-            res.render('index', { title : 'All blogs', blogs : result});
-        })
-        .catch(err => console.log(err));
-});
+// adding controllers
+// before adding controllers
+// router.get('/', (req,res) => {
+//     Blog.find().sort({ createdAt : -1 })
+//         .then(result => {
+//             res.render('index', { title : 'All blogs', blogs : result});
+//         })
+//         .catch(err => console.log(err));
+// });
+
+//after adding controllers
+router.get('/', blogController.blog_index);
+
+
 
 // Handling the post request to create a blog in the db
-router.post('/', (req,res) => {
-    const blog = Blog(req.body); 
-
-    blog.save()
-        .then(result => {
-            //logging a success record
-            console.log(`Updated ${result.title} blog successfully to the database`);
-            res.redirect('/');
-        })
-        .catch(err => console.log(err));
-});
+router.post('/', blogController.blog_create_post);
 
 // create blog form
-router.get('/create', (req,res) => {
-    res.render('create', {title : 'Create a blog'});
-});
+router.get('/create', blogController.blog_create_get);
 
 // Handling the route parameters
-router.get('/:id', (req,res) => {
-    const id = req.params.id; 
-    Blog.findById(id)
-        .then(result => {
-            res.render('details', { blog : result, title : 'Blog details' });
-        })
-        .catch(err => console.log(err));
-});
+router.get('/:id', blogController.blog_get);
 
-router.delete('/:id', (req,res) => {
-    const id = req.params.id;
-
-    Blog.findByIdAndDelete(id)
-        .then(result => {
-            res.json({ redirect : '/blogs' });
-        })
-        .catch(err => console.log(err));
-}); // NOTE: When scoping out routes if we handle requests on lets say '/blogs/create' we need set the route in request handling method as just '/create' unless if we keep it as '/blogs/create' the route we'll be using the request method for will be '/blogs/blogs/create'
+router.delete('/:id', blogController.blog_delete); // NOTE: When scoping out routes if we handle requests on lets say '/blogs/create' we need set the route in request handling method as just '/create' unless if we keep it as '/blogs/create' the route we'll be using the request method for will be '/blogs/blogs/create'
 
 // this is like a mini app()
