+++ views/blogs/create.ejs
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="en">
 
-<%- include('./partials/head.ejs') %>
+<%- include('../partials/head.ejs') %>
 
 <body>
-    <%- include('./partials/nav.ejs') %>
+    <%- include('../partials/nav.ejs') %>
 
     <div class="create-blog content">
         <!-- Creating the POST request from the form -->
@@ -20,7 +20,7 @@
         </form>
     </div>
 
-    <%- include('./partials/footer.ejs'); %>
+    <%- include('../partials/footer.ejs'); %>
 </body>
 
 </html>
