+++ views/blogs/index.ejs
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
 <html lang="en">
-<%- include('./partials/head'); %>
+<%- include('../partials/head'); %>
 <body>
     
-    <%- include('./partials/nav.ejs') // relative path to the partial %>
+    <%- include('../partials/nav.ejs') // relative path to the partial %>
 
     <div class="blogs content">
         <h2>All Blogs</h2>
@@ -24,6 +24,6 @@
         <%  } %>
     </div>
 
-    <%- include('./partials/footer.ejs'); %>
+    <%- include('../partials/footer.ejs'); %>
 </body>
 </html>
